[INTERNAL] sap.f.FlexibleColumnLayout: Add missing contextual settings propagation
Problem:
After interactive resize of the columns (by dragging the separators),
the children of the column were not always notified of the newly
updated column width. This happened because the update of the width
was not always propagated to the children: Updates were propagated
only when the interactive resize led to layout-type change, but were
not propagated on minor resizes that did not lead to layout-type change.

Solution:
Ensured that the contextual settings (a prerequisite that allows notify
childten for column resize) are propagated at the end of each resize,
regardless the impact of the resize on the current FCL layout-type.

Snippix: 487514
BCP: N/A
Change-Id: I315dcc90a96ac84b71ccf5589618bf508cf3b800