[INTERNAL] sap.uxap.ObjectPageSectionBase: introduce titleVisible
Problem: As showTitle property is not enough for user to determine
whether a title of an ObjectPageSection/SubSection is currently visible
or not, we need to expose an API in order to support the
implementation of a heading levels adjusments on FE side.

Solution: Introduce new titleVisible property on ObjectPageSectionBase
level. It is read-only and can be used for binding. The property is only
internally updated, according to the UX rules for titles visibility.

JIRA: BGSOFUIPIRIN-6182
Change-Id: Iea5fbd612b27af6dc776766a81d30550d09ab4ea