[INTERNAL] MDCTable: Fix RTA overlay
If the inner table in the "_content" aggregation has height 0px (can happen with GridTable in Auto mode), RTA incorrectly thinks that all elements inside this table are invisible, and therefore doesn't show overlays.

To fix the issue, we tell RTA to not take the domRef of whatever control is in the _content aggregation (inner table), but the one of the MDCTable itself (":sap-domref" is a magic pointer to the domRef of the "main" control).

This should not affect the size or position of any overlays, it just fixes the RTA algorithm for determining the visibility of an element.

Change-Id: Ia5d12b7974097afc3efb4266bd62f65566068652
BCP: 2370116962
CR-Id: 002075125800004713512023