[FIX] sap.f.DynamicPage: Fix RTA for content of DynamicPageTitle
Problem: When placed in the "begin" column of FlexibleColumnLayout, the
content of the title of DynamicPage is not adaptable.

Solution: The problem comes from the way RTA constructs it overlays and
the HTML structure of DynamicPage.
When RTA construct their overlays they set z-indexes, according to the
original z-indexes of HTML elements. If z-index is not available for a
certain HTML element, they search the parent tree to find the first one.
In case of FCL - begin column has z-index=3, mid column has z-index=2,
end column has z-index=1. So, as a result, all HTML elements of
DynamicPage receive z-index minimum 3 (unless they have an own one).
DynamicPageTitle already have z-index=3. However, as the Scroll
container is rendered AFTER the DynamicPageTitle, it overlaps it and
makes it unclickable --> not adaptable.
That's why now DynamicPageTitle has its own z-index=4, while the overlay
of Scroll container still receives z-index=3 (comming from begin column
of FLC, where the DynamicPage is placed). This way the title is above
the scroll container and its content is clickable.
Z-indexes of .sapFDynamicPageTitleWrapper and .sapFDynamicPageHeader are
adjusted accordingly with 1 point, in order us to keep the already
existing hierarchy of z-indexes of the HTML elements of DynamicPage.

BCP: 002075129500008317072023
Change-Id: I5ec76f250b3ebb808efff8e7b96b6de1ea7490ae
CR-Id: 002075125800005000532023
(cherry picked from commit 1904bb6ffe08b6d5f54fa45b06d9cdd044d965d2)