[FIX] sap.m.FormattedText: Allow title attribute in link
Issue:
Title attribute gets removed when passing an “a” tag to FormattedText. This is inconsistent with the use of the “tooltip” property in a Link inside the controls aggregation of FormattedText.

Solution:
Added “title” attribute for “a” tags to the list of allowed tag-property combinations that do not get automatically removed.

BCP: 2370018787
Change-Id: I095c3ad8dcb2c13e5ff966872f52f32811fcac03