[INTERNAL][FEATURE] CellSelector: Add API to retrieve selected cells
This change adds an API to retrieve the selection made with the CellSelector.

Returns the selected range.

Returns the selected cells. Instead of returning the cells directly, the selected rows and columns are returned.
This means data has to be extracted by the method caller. The method simply offers up the selection information.

The methods are still private for testing purposes and will be set to public once they were tested.

JIRA: CPOUIFTEAMB-2241
Change-Id: Id0c25ce15a88a71abcb3453729ff94434fdf0846