[INTERNAL] sap.f.FlexibleColumnLayout: refactored potential problematic method call
Concern: We are using "toReversed()" array method which is native to the
browser but only in the latest versions. There were reports where there
is an exception which breaks the whole app on older browsers which don't
support it.

Solution: Now we use "slice().reverse()" which does the same thing but
with methods widely supported in older browsers.

BCP: 2370127860
Change-Id: Icd48dfdad041cf7dc2c58534429ec215e1b5531b