[FIX] Field: Currency/unit value do not overwrite initial parts with null
A Field is bound using a Currency type and the Currency type is configured to have the currency part optional but the String type of the currency part is set to be not-nullable. The initial value of the currency is set to "". If now the users enters only a number the currency part is set to null what is not accepted for the string type.

In case only one part of an empty currency field is updated by the user, the other part must keep the initial value it gets from the model. (so no update happens on this part)

Change-Id: Iec4b6b573110274664f1bd5339c76fa7f43896f8
BCP: 2380087372