[INTERNAL] Table: Unify context menu behavior with ResponsiveTable
The beforeOpenContextMenu event is fired only before opening a custom
context menu (on non-empty standard rows).
Before, the event was fired on all rows, and even if there was no
custom context menu.

preventDefault in the beforeOpenContextMenu event does not prevent
the browser context menu.

On content rows, opening of the browser context menu is prevented only
if the table opens a context menu. It is still always prevented on
header rows, in contrast to the ResponsiveTable, because the column
header menu is opened async.

Change-Id: I635c62781a775659caebc08987e3a614aa6af4f9
Jira: CPOUIFTEAMB-2245