[FIX] sap.uxap.ObjectPageSubSection: Prevent layout flickering
Problem:
The layout flickering of the ObjectPageSubSection is caused by
misscalculating the content height, not taking into account the
additional space that is added by the CSS styling of the inner grid
inside the ObjectPageSubSection.

Solution:
This commit fixes the issue by adjusting the problematic CSS styling of
the inner grid inside the ObjectPageSubSection. This space is now
compensated by adding it to the footer spacer of the ObjectPageLayout.
This way, the content height is calculated correctly, and the layout
flickering is prevented.

BCP: 002075129500009009652023
Change-Id: Ie224116fecf96220d210a39c256d3ff0faf8be66