[FIX] sap.f.FlexibleColumnLayout: fix not visible column
Problem: When directly landing on page with TwoColumnsMidExpanded
layout, the middle column is not visible, because the pinned class is
not removed.

Solution: Pinned class is not removed, because middle column is waiting
for the resize animation of the first column to end. However, it never
ends, either an error callback is called.
It seems that, animations do not even run, when width is not specificly
set. That's why we set initial width of begin and end column to 0 in the
CSS.
Also animated class needs to be cleared when layout is changed and
before rerendering. If not cleared, error callback will be called and
animations won't run at all.

BCP: 2370126866
Change-Id: I1bc4c70333a8f73127dd7dd5049bb7d76388f2d9
CR-Id: 002075125800004999282023
(cherry picked from commit 4a16709820f3ebc0d65886d8087fa6b279363924)