[FIX] m.Table: Enter should fire selectionChange in SingleSelectMaster
 - SingleSelectMaster is a selection mode and normally only the space
event should trigger the selectionChange event but for historical
reasons the enter event should also trigger the selectionChange event.
 - Since cells are in the keyboard navigation of the Table, the
ColumnListItem delegates the enter/space events to the ListItemBase if
the focus is not on the Type/Mode controls.
 - In this event forwarding cases, the ColumnListItem#onsapenter event
handler should delegate to the ListItemBase#onsapenter, and
ListItemBase#onsapenter should not delegate back to the
ColumnListItem#onsapspace but ListItemBase#onsapspace.

BCP: 2370133922

Change-Id: I0339518fbbc5204f63c5d768d6dba1e9b8112bbc