[FIX] mdc.Table: Change default values of RowActionItem with type Navigation
If no text was set for a RowActionItem with type Navigation, the default text and icon was set by
the MDC Table. The default text was "Navigation" and the icon was "navigation-right-arrow".
But the text did not match UX specification, as the expectation is "Details".
Therefore, instead of the MDCTable defining a default, we will now use the default provided by the inner table.

BCP: 2370093621
Change-Id: I4c5baff1ef43e0c462032c8a97119a0959b6e135