[FIX] sap.m.IconTabBar: Count in filter is now visible on mobile Safari
Problem: Count is not rendering on Safari 16.6 and 17 on phone for icon tab filter in the middle of the IconTabHeader.
The issue is very specific and can not be reproduced anywhere outside of the particular use case.
It seems to be related to optimization of Safari rendering and our rendering - the text is there, but is not drawn by the browser.
After UI5 1.114 the issue is gone.

Solution: Change the positioning of the count to be relative, not absolute

CR-Id: 002075125800004981062023
BCP: 002075129500007028902023
Change-Id: Ic0b4f45a2abf13ee35ddc35893d5c05cfa177cc4