[FIX] sap.ui.unified.Menu: fix menu closing in firefox on tap
Issue: Firefox doesn't allow focus to be set on the button so focus was
set manually after a certain timeout in the change with ID: 4692376

Setting the focus back to the button will close the menu right after it
was opened.

Solution: If the menu receives a focus change event in Firefox, and the
event wasn't triggered by the items inside the menu, keep the menu open
and the focus on the correct menu item.

BCP: 2380074090
Change-Id: If646b4f1be487e3013ae8adf18369aae35f5c554