[FIX] sap.uxap.ObjectPageLayout: Downport of changes
1. sap.uxap.ObjectPageLayout: Misaligned actions are fixed

Problem: When ObjectPageLayout is in iconTabBar mode and there is only
one SubSection in a given Section, and it has some actions, the actions
overlap the content of the SubSection.

Solution: This happens because of a previous merging of the title of the
Section with the actions of the SubSection (new feature). However, in
iconTabBarMode when we do not show the title of the promoted SubSection,
the actions are overlapping the content, because the SubSection header
does not have height. Now we check for iconTabBar mode to disable the
CSS overrides for merging title with actions in this case.

BCP: 2380007216
(cherry picked from commit 27eb086f523eecb5655ede33b69fdc96c5c867dc)

2. sap.uxap.ObjectPageSubSection: Layout issue corrected

Problem:
Incorrect layout in the subSection of the page, due to non-cleared
float of an item in the subSection header

Solution:
Changed the layouting strategy in the header:
changed from use of floats to use of display: flex
(with required flex properties)

BCP: 2380131154
(cherry picked from commit 7a2d845f7522b91eae4c4a0d33c62d3aedceaee4)

Change-Id: Ibbef66a8dc7c83fe31e0b0e78fb44c2adc3f6091
CR-Id: 002075125800005066372023