[FIX] Lib/ui5loader: Provide deprecation message for pseudo-modules
With the following change/commit we introduced a regression when
requiring pseudo-modules for library types:
  Change: I8455faf59b9df581b59ffd8bc1c77904c594affd
  Commit: 453fd91cf3c7a797d5f90b531e846764276666a3

This change restores the original usage of the private "declareModule"
API of the ui5loader in Lib.js.

Furthermore the ui5loader's API is enhanced to accept a deprecation
message which is logged upon requiring the pseudo-module.

BCP: 2370138723
JIRA: CPOUI5FRAMEWORK-541
JIRA: CPOUI5FRAMEWORK-374
Change-Id: I6c414337a21d15de7e529c3658c93f0cb6232bd4