[INTERNAL] Table: Legacy-free cleanup (pointer)
During transformation to the UI5 2 build, unused dependencies are automatically removed.
This causes the TableQUnitUtils module to be removed from the Pointer QUnit test, because it is not directly used, but only indirectly through fields the util adds to the window object.

To solve this issue, one test module is refactored to directly use the util.

Change-Id: If1daf9a065342d4613a0f031c5799a140cc4f17c