[INTERNAL] sap.m.IllustratedMessage: Fixed default text console error
Problem: When we request text from the resource bundle, the
ResourceBundle file expects array at a certain func call but we pass
null and the ResourceBundle is firing error as it expects array.

Solution: We pass null in array so the console error is not printed.

Additionally some of the ".properties" files have been refactored to
exclude invalid "'(coma) character".

Fixes: https://github.com/SAP/openui5/issues/3890
BCP: 2370140729
Change-Id: I16cca8a870de9ba33956f7a068f9c06cd2288654