[FIX] sap.ui.unified.DatesRow: render week numbers
Issue:
- The week numbers weren't rendered after a regression.
A change in the primaryCalendarType property was made
so that it returns undefined in case it is not set by the user.
Therefore the week numbers weren't rendered any logner as
there is a logic to check if the primary calendar type is
gregorian.
As a result the AMP tool was flagging invalid described by
references to the week number elements.

JIRA: BGSOFUIBALKAN-6632
Change-Id: I8e532d6014b6aec142eb1f3d42ff028ec877b315