[FIX] sap.ui.unified.DatesRow: render week numbers
Issue:
- The week numbers weren't rendered after a regression.
A change in the primaryCalendarType property was made
so that it returns undefined in case it is not set by the user.
Therefore the week numbers weren't rendered any logner as
there is a logic to check if the primary calendar type is
gregorian.
As a result the AMP tool was flagging invalid described by
references to the week number elements.

Fixes: https://github.com/SAP/openui5/issues/4052

JIRA: BGSOFUIBALKAN-6632
SNOW: DINC0163770
Change-Id: I8e532d6014b6aec142eb1f3d42ff028ec877b315
CR-Id: 002075125900002523192024
(cherry picked from commit 1d659dbec26c52d98afd29b1f0841336ea07b3a7)