[FIX] sap.ui.test.action.EnterText: Revert restriction
Problem:
A restriction added with change Iecade093516c7a2f2bf944582689a604d39b2aee
prevents entering text on an non-input control. This however, causes
a backward incompatibility issue with tests that use the EnterText action
to change current value of a sap.m.Select control

Solution:
Removed the restriction, ensuring that the issue that it used to prevent
is still prevented.

BCP: 2370143073
Change-Id: Ic902043f0b1bc424200ef4ec0c2c1239cea23e6f