[INTERNAL][FIX] CellSelector: Check for existence for session object
If clearSelection is called before a session is initialized, it will fail, as session is undefined and there was no null check.

This issue has been introduced with 6008738 and is a regression that needs to be fixed in 1.120 as well.

BCP: 2370145577
CR-Id: 002075125800005320592023
Change-Id: I6e8d1ea99a7e8d39c3c8070742b46a25cd1ce757
(cherry picked from commit 4c5d19d06d4854ee76fd6ffafa62380e0a9e2b60)