[INTERNAL] sap.ui.model.odata.v2.ODataModel#(create|createEntry): remove unsupported "eTag" parameter
According to the OData V2 specification, the "If-Match" header must not
be used with POST requests for creating an entity. So "eTag" parameter
does not make sense and has to be removed.
ETag is also fetched in ODataModel#_read but never used -> remove
obsolte code too.

PS1: Remove eTag parameter for #create and #createEntry
PS2: Remove obsolete ETag code in ODataModel#_read

Change-Id: I1fedbdeeb0fa63ceac8b40e4d19c31adca9fd2d0
JIRA: CPOUI5MODELS-1079