[INTERNAL] xConfig: fix order issues in revert handling
Issue:
When changing the order of items in the xConfig, it needs to be ensured that the existing order is kept up to date for every change type. This  causes the order of columns in the table not displayed in the correct order and may lead to a mismatch in the displayed variant and expectation.

Impact:
This issue impacts the consumption of sap.m.p13n in SF and will most likely lead to customer complaints in addition.

Resolution:
Compare the updated state with the current state in the xConfig to ensure that the already created state is up to date with the already persisted state.

BCP: 2370133277
BCP: 2370127316
CR-Id: 002075125800005415062023
Change-Id: Id940ef259bc6485a1753d1f1dedcf317468c85c1
(cherry picked from commit 456f1175bd6858bec13f72167ab324300a69c109)