[INTERNAL] sap.m.IllustratedMessage: Improved resource bundle call
Problem: The 'getText' method of the ResourceBundle was previously called with an incorrect second argument. Instead of 'undefined', 'null' was being used. This could lead to potential misinterpretation of the argument by the method, leading to unexpected behavior or errors.

Solution: We've updated the call to use 'undefined' as the second argument. This ensures that the method behaves as expected, improving the reliability of the ResourceBundle calls. Although this change is not critical, it enhances the accuracy of the method invocation and ensures consistent behavior.

Change-Id: Id5f90537148969eb2fa1993bb37035536d36117a
(cherry picked from commit 877a14e8f7d80ac37d93278e18500c05518d91ef)
CR-Id: 002075125900003639882023
BCP: 2370140729