[FIX] sap.m.MenuButton: open menu when space is pressed
Issue: If a menu has been opened once already, and then closed by
switching focus to another control, the bPopupOpen flag would never be
updated. When trying to open the same menu once more with space, the
menu won't open because the menu thinks it's already open.

Solution: The sap.m.Menu and sap.unified.Menu have an isOpen() method
that checks if the menu is open instead of using boolean flags.

BCP: 2380108623
Change-Id: Ied4050658800b90acbacaf62f7e4b4f92382e6ba