[FIX] sap.m.DateTimePicker: remove explicitly setting of EN locale
Problem: When language is changed the correct translation of the DateTimePicker placeholder was not applied.

Solution: Create dedicated file for ACC / LV testing which doesn't have hard-coded locale. Rename page with the hard-coded locale so it can be clear that it is used from the visual tests and is not meant for general testing purpose.

BCP: 2380114936
BCP: 2370079317

Change-Id: Iea1fa71e017c03f741cdd2763e48d8f750dc270b