[FIX] sap.m.SinglePlanningCalendarMonthGrid: improve appointment sorting
Issue: Appointments were previously being sorted by their 'start' property which is a CalendarDate. When a JS date is converted into a CalendarDate the time part is set to the beginning of the day, so all appointments had the same time.

Solution: Sort using the correct start date found in 'data'.

JIRA: BGSOFUIBALKAN-7441
Change-Id: I78afe79ee252a5b91dae7b29c424953dc9bfbf5e