[INTERNAL][FIX] sap.ui.dt: Fix drag ghost on Safari
Previously the drag ghost on Safari was placed with fixed position as a
child of the overlay container, rendering it on top of the RTA toolbar.
For dragged items without a background this rendered artifacts from the
toolbar in the drag ghost, as the dragImage is taken directly from the
DOM.
Drag images must be visible on the DOM, thus moving it completly off
screen as it is done for the other browsers is not an option since
Safari doesn't render these sections.
A top position of 100vh seems to be a good hack, as it still considers
the item to be rendered even though it is not visible to the user.

BCP: 2380051551
Change-Id: I818ce4b6b201a183b38b05e9927373380625f08a