[INTERNAL] sap.ui.testrecorder: Transition from beforeunload event
Problem: "beforeunload" event will be soon deprecated. That's why we
need to switch to a different event, when listening for TestRecorder
window close.

Solution: Listen for "pagehide" event, which is fired when browser
window is closed, but not when it is minimized or when the tabs are
switched. TestRecorder listens for this event only when it is undocked
(opened in a separate browser window).

JIRA: BGSOFUIPIRIN-6287
Change-Id: I9b73c83dadcad5020c7e3876e2569e5808750212