[FIX] sap.uxap.ObjectPageLayout: Layout issue corrected
Problem:
Layout issue in 1.96 when performance optimization is applied
with sapUxAPObjectPageWrapperTransform
In the newer versions, the issue is no longer reproducible,
as it appears to be workarounded by the presence of another
CSS property (namely "clip-path") but the workaround may not
be fully reliable and also not possible to downport to 1.96
as it is part of a much bigger change.

Solution:
Replace the transform with will-change: scroll-position, as done
in the other scrollable page controls (sap.m.Page and sap.f.DynamicPage)

The change can then be downported to 1.96 after the required
internal-testing time in the master branch.

BCP: 002075129500009429332023
CR-Id: 002075125900003461332023
Change-Id: If9becc728df465796012135125a8112374dbeff3