[INTERNAL] sap.ui.table.Table: fix JSDoc for built-in selection APIs
Any text after a JSDoc block tag belongs to that tag and not to the
description. Therefore, the note which recommends to use the selection
plugin instead of the built-in selection APIs, must not be placed after
the @public tag, but at the end of the description. Otherwise, it
doesn't show up in the API reference.

Change-Id: I47b8518df942b5f632f96f77eba146924b6684b0