[INTERNAL] We've now enabled the property isTrustedSource to include or exclude the preview display of PDF files
If isTrustedSource is set to true, the PDFViewer opens with the displayType set to "Embedded" on desktop devices, which means that the PDF content is directly shown within the PDFViewer. Set this property to true only when the PDF is generated by the application or the PDF is scanned for viruses.

If isTrustedSource is set to false, the PDFViewer opens with the displayType set to "Link" on desktop devices, which overrides any configuration that has been provided by the application for the property displayType. This means that the PDFViewer appears as a toolbar with a download button.

BCP: 2370087602
CR-Id: 002075125900000070542024
Change-Id: Ia64c05294775e9dfc0366bcda255def88984a6d4
(cherry picked from commit c92f167655f37141fc52684e78c8a35e7d208141)