[INTERNAL] m.ListItemBase: Get the first interactive element of content
 - mdc/Table provides focusRow API with an option to put the focus to
the first interactive element.
 1. With the new keyboard handling of m/Table mode controls are in the
tab chain which was not the case before this change and the focus was
going to the first interactive cell content
 2. ui/Table does not take the selection controls into account for this
API so it also puts the focus to the first interactive cell content

Because of the above mentioned two reasons ListItemBase#getTabbables
gets a new parameter to determine whether the focus should be set in the
content (cell) area. This parameter is set to true by the
ListBase#_setFocus API

BCP: 2380133610

Change-Id: If4aff2e84714a045fc92f9b117e9ba52a49f8ae8