[FIX] sap.m.Button: adjust resource bundle API calls
Issue description:
- The sap.base.i18n.ResourceBundle.getText method takes
a string key as first argument and an array of parameter
values as second argument.
This API was used in the sap.m.Button with a string value
for the second argument. As a result a console error was thrown
with an invalid argument type message.

Solution:
- An array is provided as a second argument to the
sap.base.i18n.ResourceBundle.getText method calls
in the sap.m.Button.

BCP: 2370127868
CR-Id: 002075125800000616122024
Change-Id: Ifee39ac3e8fdcb407bad5e20a0741c3cf9050d1c