[INTERNAL] sap.tnt.ToolPage: Distance before logo image is by Horizon theme design
The distance between:
- the "Menu" button and the logo
- before the logo (if there is no "Menu" button)
is set in the Horizon less files and will come automatically for the applications.
The previous aproach (with seting it with custom class ot ID) is not good enought,
because it can't be themable.
The risk of current approach is when the elements in the ToolHeader ("Menu" button, logo, etc.)
are not ordered by the design guidelines. It is estimated as low.

BCP: 2380141775
Change-Id: I217cd32f1e45068e118f22162a9bb8ce877388c1
CR-Id: 002075125900003862822023