[FIX] sap.m.SinglePlanningCalendarMonthGrid: fixed focus on selected date or appointment
Issue: due to previous change (5338038) the default focusing
mechanism of SinglePlanningCalendarMonthGrid was broken by overriding
of applyFocusInfo method. That's why when the grid loses focus and
later there is a try to refucus again, the focus is applied only when
there is an appointment selected, otherwise previously focused date
cell wasn't focused back.

Solution: applyFocusInfo override is removed in favor of overriding
of getFocusDomRef methid, which now retuns proper focused element in
both cases (appointment or date cell selected)

BCP: 2370091991
Change-Id: I80d4da29024acd27dec14cfc31974879f90a9927