[INTERNAL] We've included the property isTrustedSource in the PDFViewer
Updating of DemoKit sample to include the property isTrustedSource.

If isTrustedSource is set to true, the PDFViewer opens with the displayType set to "Embedded" on desktop devices, which means that the PDF content is directly shown within the PDFViewer. Set this property to true only when the PDF is generated by the application or the PDF is scanned for viruses.

If isTrustedSource is set to false, the PDFViewer opens with the displayType set to "Link" on desktop devices, which overrides any configuration that has been provided by the application for the property displayType. This means that the PDFViewer appears as a toolbar with a download button.

SNOW: DINC0057645
CR-Id: 002075125900000862152024
Change-Id: Ia972d25f3a8818df00cfabe6c15471928ff7c1f9
(cherry picked from commit 7931d562a731aa47d64995e0372d8a5362141d44)