[INTERNAL] sap.uxap.LazyLoading: Improve loading of SubSections
Problem: In some situations, there is a SubSection of ObjectPageLayout,
which is not visible on the screen, but it is still loaded.

Solution: The problem happens, because the header title height and the
anchorbar height are not subtracted from the screen height in some
situations. However, it seems they should always be subtracted, in order
to make a correct calculation for the visible SubSections.

JIRA: BGSOFUIPIRIN-6373
Change-Id: I1072d48b27b24f6bc686b871ee1ace2147201613