[FIX] mTable: consider useFirstMatch on filtering existing items
If user enters a value in a Field and there are multiple entries in the ValueHelp table with this value, nothing is choosen instead of the first matching item.

If useFirstMatch is set, in case of multiple results for a user entry, the first one is used, otherwise an error is fired. If the user enters a value into the field the table of the typeahead is used to check if the value exists.
1. It checks the already loaded items of the table if there is a match.
2. If nothing is found in the existing items a backen requests is triggered to check if there a match exists.

In the first case the useFirstMatch logic is missing.

Change-Id: I28b088b197fb9113972bea6f08b9df2647f0cfa3
BCP: 002075129500010450352023
CR-Id: 002075125900003756152023
(cherry picked from commit 4f3d4a762ff021f414af1a2e45c657e7941acb83)