[INTERNAL] Prevent table export for Filter.None
When Filter.None is applied on the binding, null is
returned by Binding#getDownloadUrl. In this
scenario, the export button on the mdc.Table is disabled.

The evaluation is implemented in the
sap/m/table/Util class so that it can be used by
the SmartTable as well.

JIRA: CPOUIFTEAMB-2256
Change-Id: I65758b88bb69194fcca5b851273bc0c14b8f9e4f