[INTERNAL] table/Util: header width short circuit optimization removed
 - If the content width is larger than header length(in case of every
character has 1rem width) then column header text cannot effect the
column width calculation since in this case the column width will be
defined by the content width but this does not mean we need to do this
optimization in the dedicated "calcHeaderWidth" method e.g. SmartTable
stores the header width so that it can chop the width of the column in
case of the column should be shorter. For such cases the column header
should be calculated just just as a column header without limiting the
column header width with the content width.

BCP: 2380132964

Change-Id: I5cde7c89ccb641f94edf93ad5644f9bca8b85763