[FIX] sap.ui.core.date.UI5Date: No timeshift if given time does not exist in browser time zone
UI5Date is reusing the constructor of JavaScript Date. If the given date
and time information has to be treated as a local timestamp, then it has
to be ensured that the timestamp is treated as a local timestamp in the
given time zone.
Even if the given timestamp does not exist in the browser's time zone
due to the switch to the daylight saving time, the resulting UI5Date
instance has to return the date and time parts as given (no timeshift as
caused by the switch to DST)

PS1: Red integration test (to be run in Europe/Berlin time zone)
PS2: Rebased on top of 6026991; added further tests and a POC for fixing
     some tests
PS3: Rebased, TDD
PS4: Review comment
PS5: Resolve merge conflicts
PS6-9: Review comments (add log message for non ISO strings near switch
       to DST)

Change-Id: I8af03e0a43b0bef2ebfeed020adfcea7d01aa11c
BCP: 2370140369