[FIX] Column: Sort order if "sorted" and "sortOrder" are bound
Fixes a regression that was introduced with I32655701183b5de5bb598e66236b515f09595f32.

If both the "sorted" and "sortOrder" properties are bound, and, for example, both values in the model change from undefined to the desired value, the result is incorrect.
Because the setter for the "sorted" property also changes the "sortOrder" property for legacy support, the two-way binding of "sortOrder" changes the value in the model, before the property binding of "sortOrder" could set the property to the correct value.

Also fixes an issue of sorters not being correctly applied when the "sort" event handler did previously prevent the default.
If the default is prevented, the column internally still marks the column as sorted and tries to apply it's non-existent sorter when another column is sorted in addition.

Change-Id: Ic5ea63e75d25769826bbfb38c1ec479fd168afaa
BCP: 2380140778
Fixes: #3908