[FIX] Demokit: ApiRef now filters subclasses by visibility
Problem:
In the "Api Reference" tab, in the "View subclasses" dropdown
for an entity, the subclasses with "restricted" visibility are
displayed in the public Demokit.

Solution:
Following the existing approach for filtering the methods
by their visibility, add filtering of the subclasses by visibility
before creating the "View subclasses" dropdown.

SNOW: DINC0008990
Change-Id: I5d7de41390fdb6eca17d75363e0efe2e8a448086