[FIX] sap.f.FlexibleColumnLayout: Fixed incorrect rendering of separator
Problem: In some cases the starting drag position (once we press on it)
of the separator is incorrect.

Root Cause: Incorrect calculations which depend on the parent not having
indentation.

Solution: Change of calculations which work in both cases with or
without indentation of the parent.

Fixes: https://github.com/SAP/openui5/issues/3920
SNOW: DINC0043336
SNOW: DINC0011919
Change-Id: Ic0622ef029f71df83bcb5d7f9988592dac96536f
(cherry picked from commit 64e10bf510614712a163ace06a9a9277b073aff4)
CR-Id: 002075125900000547112024