[FIX] sap.f.DynamicPageTitle: Corrects alignment issue
Problem:
Misalignment was observed in the inner content of the
DynamicPageTitle, specifically with the GenericTag being incorrectly
positioned next to the title, and the title being not centered in
its container.

Resolution:
Refined the CSS rules applying to the dynamic title content
to resolve alignment issues. This included the removal of a few
redundant CSS rules and addition of new stronger rules.

BCP: 2370097928
Change-Id: I9ef9a96b423e8f94dc692bd90ab94135ca1bfedd