[FIX] sap.f.FlexibleColumnLayout: Documentation improved
The column separators of the FCL no longer show
navigation arrows for now, so updated the places
in the documentation where arrows are mentioned.

The parameter "isNavigationArrow" of the "stateChange"
event appears outdated, however the final spec privisions
to show navigation arrow in one case, so we can change
the name after all planned updates revisoned and finalized.

SNOW: CS20230006635616
Change-Id: Ia100b5085616fd5c6765c3798438f41221f6345f