[INTERNAL] Integration Cards: Improve API for button with single icon without text in the actionsStrip
Before when a button had an icon - it automatically was considered as sap.m.OverflowToolbarButton.
Now button with icon will be a standard sap.m.Button.
To have a sap.m.OverflowToolbarButton - the card developer can skip the "text" property.

This is more intuitive and gives more consistency between buttons and links.
It also gives more options to the developer.

Note: will not work well if text is binding, which resolves to empty string.
We have to see if we want it to work that way and if we can implement it in different change.

JIRA: BGSOFUIRODOPI-3160
Change-Id: Ia9ae6a68f5230dafd05afcd51f11574256e3b775