[INTERNAL] MockServer (tests): decouple from ODataModel V1
As preparation for UI5 2.0, tests that use ODataModel V1 either are
duplicated and adapted to ODataModel V2 (if there seems to be benefit
in keeping V1 tested) or are fully adapted to ODataModel V2.

All tests that are still based on ODataModel V1 are marked as deprecated
so that they'll be removed for 2.0.

Change-Id: Ia9a50ae3c525d9b74c3c1aa95ffe9693303908d5