[INTERNAL] sap.f.FlexibleColumnLayout: deleted volatile test
Problem: A certain spy returning different call count resulting in a
volatile test.

Root Cause: Timing issue where depending of the state of the rendering
of the FlexibleColumnLayout control and its animations, the "resume"
method of the ResizeHandler class is being triggered more than expected.

Solution: We delete the volatile code line from the test. It's
irrelevant to the test scenario. The rest of the test is testing its
purpose scenario perfectly.

SNOW: DINC0019248
Change-Id: I3bd994bfd560237948197c3c4bdf4ff4c4500495