[INTERNAL] We've now added the property isTrustedSource to include or exclude the preview display of PDF files
If isTrustedSource is set to true, the PDFViewer opens with the displayType set to "Embedded" on desktop devices, which means that the PDF content is directly shown within the PDFViewer. Set this property to true only when the PDF is generated by the application or the PDF is scanned for viruses.

If isTrustedSource is set to false, the PDFViewer opens with the displayType set to "Link" on desktop devices, which overrides any configuration that has been provided by the application for the property displayType. This means that the PDFViewer appears as a toolbar with a download button.

SNOW: DINC0022979
CR-Id: 002075125900004141692023
Change-Id: I71b2b983e8abc67de6cd679c118ed2678d227551
(cherry picked from commit 73792da055c670d8292d4b771c9ed8c9ed1f5cca)