[FEATURE]: FilterField: paste values from mdc Table
Use PasteHelper to get values from pasted texts. The PasteHelper parse function is used to get the values in the needed format.
Only the first column is used. So even if there are descriptions are provided, only keys are pasted.

Pasted into FilterField, the default operator (normally EQ) is used. The text is not checked for an operator.
Pasted into DefineConditionPanel the operator of the row is used. If BT is used and two columns are pasted, the second column is used as "to" value.
The feature to determine the operator from the pasted text was removed.

ConditionsTypeMixin created to share helper-functions between ConditionType, ConditionsType and DynamicDateRangeConditionsType.

Pasting a string containing a TAB will be interpreted as two columns and only the first value is pasted. If there is a need to support pasting of TAB this need to be supported by PasteHelper.

Change-Id: I08299b60b9c92cd384005c688dc7fa7e2906788d
JIRA: CPOUIFDENVER-2005