[INTERNAL] Column: Improve sort/filter icon updates
Fixes a performance regression that was introduced with I32655701183b5de5bb598e66236b515f09595f32

If multiple changes to the "sorted", "sortOrder", and "filtered" properties are made, the icons are updated multiple times, causing forced reflows in the browser.
This was the case already before the regression and was bad for performance. But now, for example in SmartTable scenarios, even more forced reflows are caused.

With rendering version 2, DOM patching should not be necessary anymore. The icons are now updated in the next rendering.

Change-Id: I4b239423ce1340a65070788a471d42ecd312666b
BCP: 2370135442
CR-Id: 002075125900004112402023