[FIX] sap.m.PlanningCalendar: handle date selection with min date
Issue: Date selection handler does not check if date trying to be set is
out of bounds. When selecting a year in months view, and the minimum
date is set to something after January 1st of that year, the planning
calendar will throw an exception.

Solution: Since setStartDate already handles bounds checking, focus to
the new startDate instead of January 1st.

BCP: 2370095753
Change-Id: I1e5013e658c9277aefcfecdeec99876eb218d6bb