[INTERNAL] sap.m.IllustrationPool: fixed volatile unit test
Problem: Certain test is volatile with low success rate.

Root Cause: We are expecting a stub to be called once but due to the
nature of the async execution, sometimes it's also called by other test
in the meantime.

Solution: Move the problematic test after the volatile test in the same
module to ensure execution after proper calculations of the no longer
volatile test.

SNOW: DINC0021870
Change-Id: I85d957ae4d2baee51124e084c011c080dbca3cae