[INTERNAL][FIX] SimpleForm: fix performance in AddSimpleFormField
A recent bug fix introduced a performance issue when adding many fields to a SimpleForm.
This new change fixes the original issue and also does not impact the performance as much.

A new modifier function was introduced: 'recreateAggregation', which is used in the changeHandler.
This function first gets the aggregation (async) to then synchronously
remove the whole content and adds the passed content to the aggregation.
By not adding the content temporarily to the dependends aggregation the performance issue should be gone.

SNOW: CS20230006394894
CR-Id: 002075125900004274382023
Change-Id: I439df807653bdfee78482926de27811c37e0630a
(cherry picked from commit 99f4556d53f30a083b7fb91a72f9af93f691032c)