[INTERNAL] sap.ui.documentation (tests): avoid absolute URL paths
Absolute URL paths don't work in all use cases. They either refer to the
local development setup (e.g. /resources/something) or to the J2EE setup
as used in the nightly build or infra voter (/qunit-tests/resources/...)

Instead of absolute paths, relative paths should be used. If a <base> is
set, it should be set relative to the page location. All other paths
then should be specified relative to the <base>.

Change-Id: I77522e1660571a5c273c5a16cf286abf91e5d75a