[INTERNAL] Demo Kit: Update Mock API index
This is a follow-up to Ie39e5bf436b23af3271509b1e41de993d4287089 (commit
e1c5e067926f2ae898a8f196dec133150696b278).

The mock API index was updated to reflect the changes in the API index
of the UI5 SDK. This is necessary to make the integration tests work as
expected.

Jira: BGSOFUIPIRIN-6362
Change-Id: I90be880b67af599fb7c3d0c77f00f8380e7899d4