[INTERNAL] sap.f.FlexibleColumnLayout: Update contextual settings on live resize
Problem:
When the FCL contains a sap.m.Table with autoPopinMode="true",
its columns are not always updated during live-resize of the FCL
column [by dragging the column separators], but only when the
column layout-type changed during drag or after the user released
the mouse.

The observed behaviour was intentional for performance reasons.
However, we found that the existing implementation inside
sap.ui.core.Element allows the controls inside the FCL columns
to be notified in time without performance drawback:

If the controls nested inside the FCL column subscribe to be
notified for updates only when the width reached cerain breakpoints
(as the sap.m.Table columns currently do), then it seems safe for
the FCL to propagate the updates to the width at each tiny resize
(as sap.ui.core.Element will notify the underlying controls only if
the new width passed the required breakpoints).

Solution:
1) FCL now always updates the column contextual width (on each
small resize during interactive resize with drag&drop)
2) ObjectPage and DynamicPage now subscribe to be notified for update
of the contextual width only when the width passed the pre-registered
breakpoints.

Change-Id: Ib8ec9abe6752761bcf92dc4c5a6b584c13fa680d