[FIX] sap.m.ProgressIndicator: Fiori 3 theme alignment
Problem:
The ProgressIndicator margin is not correctly applied in
different content densities across the Fiori 3 theme. That leads to
problems regarding this control when used inside a Table. Because of the
incorrect margin, the ProgressIndicator occupying space is increased
which leads to increasing the row height in the Table. This affects the
scrolling behaviour of the Table.

Solution:
The ProgressIndicator margin is now correctly applied in different
content densities across the Fiori 3 theme, following the design
guidelines.

SNOW: CS20230006501424
CR-Id: 002075125900000108002024
Change-Id: Ie938c3d0f76041ec0868bbcac682e568fc52b8f8
(cherry picked from commit 913dd95fca3e7ee5ac45fedfb4eea368c6bc0555)